Refactor tests to use create_replication_client for improved client management
